+++ nebula/AsyncClient.py
@@ -8,11 +8,11 @@
 import asyncio
 
 from .Common import *
-from thrift.transport import TSocket
-from thrift.transport import TTransport
-from thrift.protocol import TBinaryProtocol
+from nebula.thrift.transport import TSocket
+from nebula.thrift.transport import TTransport
+from nebula.thrift.protocol import TBinaryProtocol
 
-from graph import AsyncGraphService
+from nebula.graph import AsyncGraphService
 
 
 class AsyncGraphClient(object):