+++ nebula/graph/AsyncGraphService.py
@@ -33,6 +33,13 @@ if not '__pypy__' in sys.builtin_module_names:
 all_structs = []
 UTF8STRINGS = bool(0) or sys.version_info.major >= 3
 
+import thrift
+if six.PY3 and not thrift.trollius:
+  import asyncio
+  from thrift.util.asyncio import call_as_future
+else:
+  import trollius as asyncio
+  from thrift.util.trollius import call_as_future
 from thrift.util.Decorators import *
 
 class Iface:
@@ -632,19 +639,11 @@ execute_result.__getstate__ = lambda self: self.__dict__.copy()
 execute_result.__setstate__ = execute_result__setstate__
 
 class Client(Iface):
-  def __enter__(self):
-    return self
-
-  def __exit__(self, type, value, tb):
-    self._iprot.trans.close()
-    if self._iprot is not self._oprot:
-      self._oprot.trans.close()
-
-  def __init__(self, iprot, oprot=None):
-    self._iprot = self._oprot = iprot
-    if oprot != None:
-      self._oprot = oprot
+  def __init__(self, oprot, loop=None):
+    self._oprot = oprot
+    self._loop = loop or asyncio.get_event_loop()
     self._seqid = 0
+    self._futures = {}
 
   def authenticate(self, username=None, password=None):
     """
@@ -652,8 +651,10 @@ class Client(Iface):
      - username
      - password
     """
+    self._seqid += 1
+    fut = self._futures[self._seqid] = asyncio.Future(loop=self._loop)
     self.send_authenticate(username, password)
-    return self.recv_authenticate()
+    return fut
 
   def send_authenticate(self, username=None, password=None):
     self._oprot.writeMessageBegin('authenticate', TMessageType.CALL, self._seqid)
@@ -664,26 +665,40 @@ class Client(Iface):
     self._oprot.writeMessageEnd()
     self._oprot.trans.flush()
 
-  def recv_authenticate(self, ):
-    (fname, mtype, rseqid) = self._iprot.readMessageBegin()
+  def recv_authenticate(self, iprot, mtype, rseqid):
+    try:
+        fut = self._futures.pop(rseqid)
+    except KeyError:
+        return   # request timed out
     if mtype == TMessageType.EXCEPTION:
       x = TApplicationException()
-      x.read(self._iprot)
-      self._iprot.readMessageEnd()
-      raise x
+      x.read(iprot)
+      iprot.readMessageEnd()
+      fut.set_exception(x)
+      return
     result = authenticate_result()
-    result.read(self._iprot)
-    self._iprot.readMessageEnd()
+    try:
+      result.read(iprot)
+    except Exception as e:
+      fut.set_exception(e)
+      return
+    iprot.readMessageEnd()
     if result.success != None:
-      return result.success
-    raise TApplicationException(TApplicationException.MISSING_RESULT, "authenticate failed: unknown result");
+      fut.set_result(result.success)
+      return
+    fut.set_exception(TApplicationException(TApplicationException.MISSING_RESULT, "authenticate failed: unknown result"))
+    return
 
   def signout(self, sessionId=None):
     """
     Parameters:
      - sessionId
     """
+    self._seqid += 1
+    fut = self._futures[self._seqid] = asyncio.Future(loop=self._loop)
     self.send_signout(sessionId)
+    fut.set_result(None)
+    return fut
 
   def send_signout(self, sessionId=None):
     self._oprot.writeMessageBegin('signout', TMessageType.CALL, self._seqid)
@@ -698,8 +713,10 @@ class Client(Iface):
      - sessionId
      - stmt
     """
+    self._seqid += 1
+    fut = self._futures[self._seqid] = asyncio.Future(loop=self._loop)
     self.send_execute(sessionId, stmt)
-    return self.recv_execute()
+    return fut
 
   def send_execute(self, sessionId=None, stmt=None):
     self._oprot.writeMessageBegin('execute', TMessageType.CALL, self._seqid)
@@ -710,27 +727,38 @@ class Client(Iface):
     self._oprot.writeMessageEnd()
     self._oprot.trans.flush()
 
-  def recv_execute(self, ):
-    (fname, mtype, rseqid) = self._iprot.readMessageBegin()
+  def recv_execute(self, iprot, mtype, rseqid):
+    try:
+        fut = self._futures.pop(rseqid)
+    except KeyError:
+        return   # request timed out
     if mtype == TMessageType.EXCEPTION:
       x = TApplicationException()
-      x.read(self._iprot)
-      self._iprot.readMessageEnd()
-      raise x
+      x.read(iprot)
+      iprot.readMessageEnd()
+      fut.set_exception(x)
+      return
     result = execute_result()
-    result.read(self._iprot)
-    self._iprot.readMessageEnd()
+    try:
+      result.read(iprot)
+    except Exception as e:
+      fut.set_exception(e)
+      return
+    iprot.readMessageEnd()
     if result.success != None:
-      return result.success
-    raise TApplicationException(TApplicationException.MISSING_RESULT, "execute failed: unknown result");
+      fut.set_result(result.success)
+      return
+    fut.set_exception(TApplicationException(TApplicationException.MISSING_RESULT, "execute failed: unknown result"))
+    return
 
 
 class Processor(Iface, TProcessor):
   _onewayMethods = ("signout",)
 
-  def __init__(self, handler):
+  def __init__(self, handler, loop=None):
     TProcessor.__init__(self)
     self._handler = handler
+    self._loop = loop or asyncio.get_event_loop()
     self._processMap = {}
     self._priorityMap = {}
     self._processMap["authenticate"] = Processor.process_authenticate
@@ -745,48 +773,46 @@ class Processor(Iface, TProcessor):
     l.extend(Processor._onewayMethods)
     return tuple(l)
 
-  @process_main()
+  @process_main(asyncio=True)
   def process(self,): pass
 
-  @process_method(authenticate_args, oneway=False)
-  def process_authenticate(self, args, handler_ctx):
+  @process_method(authenticate_args, oneway=False, asyncio=True)
+  def process_authenticate(self, args, handler_ctx, seqid, oprot, fn_name):
     result = authenticate_result()
-    try:
-      result.success = self._handler.authenticate(args.username, args.password)
-    except:
-      ex = sys.exc_info()[1]
-      self._event_handler.handlerError(handler_ctx, 'authenticate', ex)
-      result = Thrift.TApplicationException(message=str(ex))
-    return result
+    if should_run_on_thread(self._handler.authenticate):
+      fut = self._loop.run_in_executor(None, self._handler.authenticate, args.username, args.password)
+    else:
+      fut = call_as_future(self._handler.authenticate, self._loop, args.username, args.password)
+    fut.add_done_callback(lambda f: write_results_after_future(result, self._event_handler, handler_ctx, seqid, oprot, fn_name, {}, f))
+    return fut
 
-  @process_method(signout_args, oneway=True)
-  def process_signout(self, args, handler_ctx):
-    try:
-      self._handler.signout(args.sessionId)
-    except:
-      ex = sys.exc_info()[1]
-      self._event_handler.handlerError(handler_ctx, 'signout', ex)
-      result = Thrift.TApplicationException(message=str(ex))
+  @process_method(signout_args, oneway=True, asyncio=True)
+  def process_signout(self, args, handler_ctx, seqid, oprot, fn_name):
+    if should_run_on_thread(self._handler.signout):
+      fut = self._loop.run_in_executor(None, self._handler.signout, args.sessionId)
+    else:
+      fut = call_as_future(self._handler.signout, self._loop, args.sessionId)
+    return fut
 
-  @process_method(execute_args, oneway=False)
-  def process_execute(self, args, handler_ctx):
+  @process_method(execute_args, oneway=False, asyncio=True)
+  def process_execute(self, args, handler_ctx, seqid, oprot, fn_name):
     result = execute_result()
-    try:
-      result.success = self._handler.execute(args.sessionId, args.stmt)
-    except:
-      ex = sys.exc_info()[1]
-      self._event_handler.handlerError(handler_ctx, 'execute', ex)
-      result = Thrift.TApplicationException(message=str(ex))
-    return result
+    if should_run_on_thread(self._handler.execute):
+      fut = self._loop.run_in_executor(None, self._handler.execute, args.sessionId, args.stmt)
+    else:
+      fut = call_as_future(self._handler.execute, self._loop, args.sessionId, args.stmt)
+    fut.add_done_callback(lambda f: write_results_after_future(result, self._event_handler, handler_ctx, seqid, oprot, fn_name, {}, f))
+    return fut
 
 Iface._processor_type = Processor
 
 class ContextProcessor(ContextIface, TProcessor):
   _onewayMethods = ("signout",)
 
-  def __init__(self, handler):
+  def __init__(self, handler, loop=None):
     TProcessor.__init__(self)
     self._handler = handler
+    self._loop = loop or asyncio.get_event_loop()
     self._processMap = {}
     self._priorityMap = {}
     self._processMap["authenticate"] = ContextProcessor.process_authenticate
@@ -801,42 +827,40 @@ class ContextProcessor(ContextIface, TProcessor):
     l.extend(ContextProcessor._onewayMethods)
     return tuple(l)
 
-  @process_main()
+  @process_main(asyncio=True)
   def process(self,): pass
 
-  @process_method(authenticate_args, oneway=False)
-  def process_authenticate(self, args, handler_ctx):
+  @process_method(authenticate_args, oneway=False, asyncio=True)
+  def process_authenticate(self, args, handler_ctx, seqid, oprot, fn_name):
     result = authenticate_result()
-    try:
-      result.success = self._handler.authenticate(handler_ctx, args.username, args.password)
-    except:
-      ex = sys.exc_info()[1]
-      self._event_handler.handlerError(handler_ctx, 'authenticate', ex)
-      result = Thrift.TApplicationException(message=str(ex))
-    return result
+    if should_run_on_thread(self._handler.authenticate):
+      fut = self._loop.run_in_executor(None, self._handler.authenticate, handler_ctx, args.username, args.password)
+    else:
+      fut = call_as_future(self._handler.authenticate, self._loop, handler_ctx, args.username, args.password)
+    fut.add_done_callback(lambda f: write_results_after_future(result, self._event_handler, handler_ctx, seqid, oprot, fn_name, {}, f))
+    return fut
 
-  @process_method(signout_args, oneway=True)
-  def process_signout(self, args, handler_ctx):
-    try:
-      self._handler.signout(handler_ctx, args.sessionId)
-    except:
-      ex = sys.exc_info()[1]
-      self._event_handler.handlerError(handler_ctx, 'signout', ex)
-      result = Thrift.TApplicationException(message=str(ex))
+  @process_method(signout_args, oneway=True, asyncio=True)
+  def process_signout(self, args, handler_ctx, seqid, oprot, fn_name):
+    if should_run_on_thread(self._handler.signout):
+      fut = self._loop.run_in_executor(None, self._handler.signout, handler_ctx, args.sessionId)
+    else:
+      fut = call_as_future(self._handler.signout, self._loop, handler_ctx, args.sessionId)
+    return fut
 
-  @process_method(execute_args, oneway=False)
-  def process_execute(self, args, handler_ctx):
+  @process_method(execute_args, oneway=False, asyncio=True)
+  def process_execute(self, args, handler_ctx, seqid, oprot, fn_name):
     result = execute_result()
-    try:
-      result.success = self._handler.execute(handler_ctx, args.sessionId, args.stmt)
-    except:
-      ex = sys.exc_info()[1]
-      self._event_handler.handlerError(handler_ctx, 'execute', ex)
-      result = Thrift.TApplicationException(message=str(ex))
-    return result
+    if should_run_on_thread(self._handler.execute):
+      fut = self._loop.run_in_executor(None, self._handler.execute, handler_ctx, args.sessionId, args.stmt)
+    else:
+      fut = call_as_future(self._handler.execute, self._loop, handler_ctx, args.sessionId, args.stmt)
+    fut.add_done_callback(lambda f: write_results_after_future(result, self._event_handler, handler_ctx, seqid, oprot, fn_name, {}, f))
+    return fut
 
 ContextIface._processor_type = ContextProcessor
 
 fix_spec(all_structs)
 del all_structs
+