+++ nebula/ngData/data.py
@@ -5,7 +5,7 @@
 # This source code is licensed under Apache 2.0 License,
 # attached with Common Clause Condition 1.0, found in the LICENSES directory.
 
-from common.ttypes import SupportedType
+from nebula.common.ttypes import SupportedType
 import struct
 import six
 from datetime import datetime